+++ lib/force_mvc.dart
@@ -25,7 +25,7 @@ part 'server/force_handler_interceptor.dart';
 part 'server/force_interceptors_collection.dart';
 part 'server/force_path_analyzer.dart';
 part 'server/force_serving_files.dart';
-
+part 'server/force_response_hooks.dart';
 // security
 part 'security/security_context_holder.dart';
 part 'security/security_strategy.dart';