+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file contains highlights of what changes on each version of the forcemvc pa
 #### Pub version 0.3.3 ####
 
 - Error logging when the server don't start.
+- Add responseHooks into the web_server.
 
 #### Pub version 0.3.2+1 & 0.3.2+2 ####